+++ src/browser/mac.rs
@@ -1,11 +1,13 @@
+use smol_str::SmolStr;
+
 pub trait MacSign {
-    fn sign(_message: &[u8]) -> Option<String> {
+    fn sign(_message: &[u8]) -> Option<SmolStr> {
         None
     }
 }
 
 pub trait MacVerify {
-    fn verify(_message: &[u8], _signature: Option<&str>) -> Result<bool, String> {
+    fn verify(_message: &[u8], _signature: Option<&str>) -> Result<bool, SmolStr> {
         Ok(true)
     }
 }