+++ CHANGELOG.md
@@ -4,11 +4,27 @@ All notable changes to this project will be documented in this file. The
 format is based on [Keep a Changelog], and this project aims to follow
 [Semantic Versioning].
 
-## [0.3.2] - 2024-10-07
+## [0.5.0] - 2024-10-30
+
+### Changed
+
+- `String` usage replaced with `SmolStr`
+
+- `uformat` replaced with `uformat_smolstr`
+
+- bump up dependencies versions
+
+## [0.4.2] - 2024-10-07
 
 ### Changed
 
 - bump up dependencies versions
+
+## [0.4.1] - 2024-08-31
+
+### Fixed
+
+- versions 0.3.x were wrong, they changed interface
 
 ## [0.3.1] - 2024-08-31